% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_mfd.R
\name{basis_mfd}
\alias{basis_mfd}
\title{Make basis expansion matrix from multivariate functional data}
\usage{
basis_mfd(
  X,
  grid = NULL,
  basis = "bspline",
  n_basis = 4,
  n_order = 4,
  FVE = 0.9,
  select_K = "n_basis",
  gram = TRUE
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{grid}{a vector containing m timepoints}

\item{basis}{a choice of the basis. "bspline" (B-spline), "ufpca" (uFPCA) and "fpca_ps" (FPCA under partial separability) are supported.}

\item{n_basis}{the number of basis for the B-spline basis expansion}

\item{n_order}{the order of B-spline basis functions. Default is 4 (cubic B-spline)}

\item{FVE}{Fraction of variance explained (Default is 0.90)}

\item{select_K}{the number of FPCs. "FVE" and "n_basis" are supported. (Default is selected by "n_basis")}

\item{gram}{If TRUE, Gram-Schmidt orthogonalization is performed for the estimated basis coefficients. (Default is TRUE)}
}
\value{
a \code{basis_mfd} object
}
\description{
Make basis expansion matrix from multivariate functional data
}
