% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspt_hfd.R
\name{dspt_hfd}
\alias{dspt_hfd}
\title{Data-Splitting Projection Test for High-Dimensional Functional Data}
\usage{
dspt_hfd(
  X1,
  X2,
  penalty = "scad",
  split_prop = 0.5,
  tuning = FALSE,
  tune_method = "cv",
  n_basis = 4,
  lambda = 0.1,
  n_cores = 1,
  ...
)
}
\arguments{
\item{X1}{a n1-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{X2}{a n2-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{penalty}{the default penalty type is "scad" (group SCAD). ("l1" is also supported which is the \code{hdflda} direction)}

\item{split_prop}{the proportion of data-splitting (split_prop for estimating the optimal projection direction; 1-split_prop for test). Default is 0.5.}

\item{tuning}{If \code{tuning = TRUE}, the parameter tuning is performed. (\code{n_basis}, \code{lambda})}

\item{tune_method}{"cv", "abic" and "abic-aic" are supported. Default is "cv".}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}

\item{lambda}{a penalty parameter for penalty function}

\item{n_cores}{a number cores for parallel computing}

\item{...}{additional parameters or \code{opt_proj_dir} ("scad") or \code{hdflda} ("l1")}
}
\value{
a list with class \code{htest} containing the following components:
\item{statistic}{the value of the t-statistic from the projected samples}
\item{p.value}{the p-value for the test}
\item{method}{a character string indicating what type of t-test was performed}
\item{data.name}{a character string giving the name(s) of the data}
\item{obj_test}{the object from \code{t.test()}}
\item{obj_proj}{the \code{opt_proj_dir} object}
\item{obj_proj_tune}{the \code{tune.opt_proj_dir} object if \code{tuning = TRUE}}
}
\description{
Two-sample test for high-dimensional functional means via data-splitting projection test onto the optimal projection direction
The optimal projection direction is corresponding to the high-dimensional functional linear discriminant direction under the sparsity assumption.
The optimal projection direction is estimated by the block coordinate descent (BCD) algorithm with local quadratic approximation (LQA).
}
\references{
Kim, H. and Park, J. (2026+). Two-sample Projection Test for High-Dimensional Functional Data, Submitted.
}
