% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fglm.R
\name{fglm}
\alias{fglm}
\title{Functional generalized linear model (FGLM)}
\usage{
fglm(
  X,
  y,
  family = "binomial",
  grid = NULL,
  basis = "bspline",
  FVE = 0.9,
  K = NULL,
  n_basis = 20
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{family}{the parameter of family for \code{glm()}. Default is "binomial".}

\item{grid}{a vector containing m timepoints}

\item{basis}{"fpca" (FPCA) or "bspline" (B-spline). Default is "bspline".}

\item{FVE}{the fraction of variance explained to choose the number of the FPCs}

\item{K}{the number of FPCs}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}
}
\value{
a \code{fglm} object
}
\description{
Scalar on function generalized linear model
}
