% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso.R
\name{flasso}
\alias{flasso}
\title{Lasso for functional generalized linear model based on group lasso}
\usage{
flasso(
  X,
  y,
  family = "binomial",
  grid = NULL,
  basis = "bspline",
  lambda = 0.1,
  alpha = 0.05,
  FVE = 0.9,
  K = NULL,
  n_basis = 20
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{family}{the parameter of family for \code{glm()}. Default is "binomial".}

\item{grid}{a vector containing m timepoints}

\item{basis}{"fpca" (FPCA) or "bspline" (B-spline). Default is "bspline".}

\item{lambda}{a penalty parameter for group lasso. Default is 0.1. If it is \code{NULL} or the vector containing the candidates of \code{lambda}, cross-validation is performed.}

\item{alpha}{a relative weight for L1-regularization (1-alpha for L2-regularization). Default is 0.05.}

\item{FVE}{the fraction of variance explained to choose the number of the FPCs}

\item{K}{the number of FPCs}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots. If it is \code{NULL} or the vector containing the candidates of \code{n_basis}, cross-validation is performed.}
}
\value{
a \code{flasso} object
}
\description{
Scalar on function generalized linear model
It implements the group lasso for FPC scores or B-spline bases coefficients.
}
