% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfd_proj_test.R
\name{hdfd_proj_test}
\alias{hdfd_proj_test}
\title{Two-sample Projection Test for High-Dimensional Functional Data}
\usage{
hdfd_proj_test(
  X1,
  X2,
  penalty = "scad",
  split_prop = 0.5,
  tuning = FALSE,
  tune_method = "cv",
  n_basis = 4,
  lambda = 0.1,
  n_cores = 1,
  ...
)
}
\arguments{
\item{X1}{a n1-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{X2}{a n2-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{penalty}{the default penalty type is "scad" (group SCAD). ("l1" is also supported which is the \code{hdflda} direction)}

\item{split_prop}{the proportion of data-splitting (split_prop for estimating the optimal projection direction; 1-split_prop for test). Default is 0.5.}

\item{tuning}{If \code{tuning = TRUE}, the parameter tuning is performed. (\code{n_basis}, \code{lambda})}

\item{tune_method}{"cv", "abic" and "abic-aic" are supported. Default is "cv".}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}

\item{lambda}{a penalty parameter for penalty function}

\item{n_cores}{a number cores for parallel computing}

\item{...}{additional parameters or \code{opt_proj_dir} ("scad") or \code{hdflda} ("l1")}
}
\value{
a \code{htest} object
}
\description{
The data-splitting projection test onto the optimal projection direction for high-dimensional functional data
The optimal projection direction is corresponding to the high-dimensional functional linear discriminant direction under the sparsity assumption.
The optimal projection direction is estimated by the block coordinate descent (BCD) algorithm with local quadratic approximation (LQA).
}
\references{
Kim, H. and Park, J. (2025+). Two-sample Projection Test for High-Dimensional Functional Data, Submitted.
}
