% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdflda.R
\name{hdflda}
\alias{hdflda}
\title{LDA for high-dimensional functional data}
\usage{
hdflda(
  X,
  y,
  grid = NULL,
  basis = "bspline",
  n_basis = 4,
  lambda = 0.1,
  tol = 1e-07
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{grid}{a vector containing m timepoints}

\item{basis}{"bspline" is only supported}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}

\item{lambda}{a penalty parameter for L1-regularization}

\item{tol}{a tolerance rate to define the sparse discriminant set}
}
\value{
a \code{hdflda} object
}
\description{
Linear discriminant analysis for high-dimensional functional data
}
\references{
Xue, K., Yang, J., & Yao, F. (2023). Optimal linear discriminant analysis for high-dimensional functional data. Journal of the American Statistical Association, 1-10.
}
