% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdflr.R
\name{hdflr}
\alias{hdflr}
\title{Functional linear regression for high-dimensional functional data with group SCAD penalty}
\usage{
hdflr(
  X,
  y,
  grid = NULL,
  penalty = "gr_scad",
  type = "regress",
  basis = "bspline",
  n_basis = 4,
  lambda = 0.1,
  a = 3.7,
  max_iter = 1000,
  tol = 1e-07
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{grid}{a vector containing m timepoints}

\item{penalty}{the penalty type ("gr_scad" is only supported)}

\item{type}{"regress" for regression and "classif" for LDA direction estimation}

\item{basis}{"bspline" is only supported}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}

\item{lambda}{a penalty parameter for L1-regularization}

\item{a}{a parameter for SCAD penalty}

\item{max_iter}{a maximum iteration number of the coordinate descent algorithm}

\item{tol}{a tolerance limit of the coordinate descent algorithm}
}
\value{
a \code{hdflr} object
}
\description{
Scalar on function linear regression with group SCAD penalty in high-dimension
}
