% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis_mf.R
\name{make_basis_mf}
\alias{make_basis_mf}
\title{Make basis expansion matrix from multivariate functional data}
\usage{
make_basis_mf(
  X,
  grid = NULL,
  basis = "fpca",
  FVE = 0.9,
  K = NULL,
  n_knots = 20,
  gram = F
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{grid}{a vector containing m timepoints}

\item{basis}{a choice of the basis. "fpca" (FPCA) or "bspline" (B-spline) is supported.}

\item{FVE}{Fraction of variance explained (Default is 0.90)}

\item{K}{the number of FPCs (Default is selected by FVE)}

\item{n_knots}{the number of knots for the cubic B-spline basis expansion}

\item{gram}{If TRUE, Gram-Schmidt orthogonalization is performed for the estimated basis coefficients. (Default is FALSE)}
}
\value{
a \code{make_basis_mf} object
}
\description{
Make basis expansion matrix from multivariate functional data
}
