% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_mfd.R
\name{predict.basis_mfd}
\alias{predict.basis_mfd}
\title{Predict the basis coefficient matrix from new multivariate functional data}
\usage{
\method{predict}{basis_mfd}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{basis_mfd} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{...}{Not used}
}
\value{
a basis coefficient matrix (n-(K_1 + ... + K_p) matrix)
}
\description{
Predict the basis coefficient matrix from new multivariate functional data
}
