% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fglm.R
\name{predict.fglm}
\alias{predict.fglm}
\title{Predict the class labels using the FGLM}
\usage{
\method{predict}{fglm}(object, newdata, threshold = 0.5, ...)
}
\arguments{
\item{object}{a \code{fglm} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{threshold}{the cutoff value for the binary classification. Default is 0.5.}

\item{...}{Not used}
}
\value{
a predicted class
}
\description{
Obtain the predicted class labels using \code{fglm} object
}
