% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso.R
\name{predict.flasso}
\alias{predict.flasso}
\title{Predict the class labels using \code{flasso}}
\usage{
\method{predict}{flasso}(object, newdata, threshold = 0.5, ...)
}
\arguments{
\item{object}{a \code{flasso} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{threshold}{the cutoff value for the binary classification. Default is 0.5.}

\item{...}{Not used}
}
\value{
a predicted class
}
\description{
Obtain the predicted class labels using \code{flasso} object
}
