% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdflda.R
\name{predict.hdflda}
\alias{predict.hdflda}
\title{Predict the class labels using the high-dimensional functional LDA}
\usage{
\method{predict}{hdflda}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{hdflda} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{...}{Not used}
}
\value{
a \code{hdflda} object
}
\description{
Obtain the predicted class labels using \code{hdflda} object
}
\references{
Xue, K., Yang, J., & Yao, F. (2023). Optimal linear discriminant analysis for high-dimensional functional data. Journal of the American Statistical Association, 1-10.
}
