% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdflr.R
\name{predict.hdflr}
\alias{predict.hdflr}
\title{Predict the new data using the high-dimensional functional linear regression}
\usage{
\method{predict}{hdflr}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{hdflr} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{...}{Not used}
}
\value{
a \code{hdflr} object
}
\description{
Obtain the prediction using \code{hdflr} object
}
