% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis_mf.R
\name{predict.make_basis_mf}
\alias{predict.make_basis_mf}
\title{Predict the basis coefficient matrix from new multivariate functional data}
\usage{
\method{predict}{make_basis_mf}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{make_basis_mf} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{...}{Not used}
}
\value{
a basis coefficient matrix (n-(K_1 + ... + K_p) matrix)
}
\description{
Predict the basis coefficient matrix from new multivariate functional data
}
