% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_proj_dir.R
\name{predict.opt_proj_dir}
\alias{predict.opt_proj_dir}
\title{Predict the class of new data from \code{opt_proj_dir} object}
\usage{
\method{predict}{opt_proj_dir}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{opt_proj_dir} object}

\item{newdata}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{...}{Not used}
}
\value{
a \code{opt_proj_dir} object
}
\description{
The prediction is obtained using the high-dimensional LDA direction with group non-convex penalty (i.e., optimal projection direction) under the Gaussian assumption
}
