% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCA.R
\name{predict.uFPCA}
\alias{predict.uFPCA}
\title{Predict using \code{uFPCA} object}
\usage{
\method{predict}{uFPCA}(object, newdata, K = NULL, type = "fpc.score", ...)
}
\arguments{
\item{object}{a \code{uFPCA} object}

\item{newdata}{a n-m matrix (n curves observed from m timepoints)}

\item{K}{the number of FPCs (Default is selected by FVE)}

\item{type}{"fpc.score" (FPC score prediction) or "raw" (curve prediction) (Default is "fpc.score")}

\item{...}{Not used}
}
\value{
a \code{uFPCA} object
}
\description{
Get the predicted FPC scores or the curve prediction using \code{uFPCA} object
}
