% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_proj_dir.R
\name{tune.opt_proj_dir}
\alias{tune.opt_proj_dir}
\title{Tuning Parameter election for the Estimation of Optimal Projection Direction for High-dimensional Functional Data}
\usage{
tune.opt_proj_dir(
  X,
  y,
  tune_method = "cv",
  grid = NULL,
  penalty = "scad",
  basis = "bspline",
  n_basis_list = NULL,
  lambda_list = NULL,
  n_cores = 1,
  measure = "accuracy",
  K = 5,
  tie_break = "sparse",
  ...
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{tune_method}{"cv", "abic" or "abic-aic" (default is "cv")}

\item{grid}{a vector containing m timepoints}

\item{penalty}{the penalty type ("scad" is only supported)}

\item{basis}{Default is "bspline" ("fpca" is possible but takes huge times if p is high)}

\item{n_basis_list}{a vector containing the candidate of \code{n_basis} (the number of cubic B-spline bases using \code{n_basis}-2 knots)}

\item{lambda_list}{a vector containing the candidate of \code{lambda} (a penalty parameter for L1-regularization)}

\item{n_cores}{a number cores for parallel computing}

\item{measure}{the measure for the K-fold cross-validation. "accuracy", "cross.entropy", "mahalanobis" (Default is "accuracy")}

\item{K}{the number of folds for K-fold CV}

\item{tie_break}{the tie breaking rule for cross-validation. "sparse"(default) choose the largest \code{lambda} and the smallest \code{n_basis}; "random" choose randomly}

\item{...}{additional parameters for \code{opt_proj_dir}}
}
\value{
a \code{tune.opt_proj_dir} object
}
\description{
Tuning parameter (lambda, n_basis) selection for \code{opt_proj_dir}
}
\details{
Select the optimal \code{n_basis} and \code{lambda} for \code{opt_proj_dir} using K-fold cross-validation with accuracy and Mahalanobis distance or ABIC, ABIC-AIC.
Default tuning method is K-fold cross-validation with a measure classification error rate.
Parallel computing can be used by using the \code{doParallel} package usages.
}
\references{
Kim, H. and Park, J. (2026+). Two-sample Projection Test for High-Dimensional Functional Data, Submitted.
}
