% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFPCA.R
\name{uFPCA}
\alias{uFPCA}
\title{Univariate FPCA}
\usage{
uFPCA(X, grid = NULL, K = NULL, FVE = 0.95, centered = FALSE)
}
\arguments{
\item{X}{a n-m matrix (n curves observed from m timepoints)}

\item{grid}{a vector containing m timepoints}

\item{K}{the number of FPCs (Default is selected by FVE)}

\item{FVE}{Fraction of variance explained (Default is 0.95)}

\item{centered}{If TRUE, centering of curves is not conducted. (Default is FALSE)}
}
\value{
a \code{uFPCA} object
}
\description{
Functional Principal Component Analysis for Univariate Functional Data
}
