% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_mfd.R
\name{basis_mfd}
\alias{basis_mfd}
\title{Make basis expansion matrix from multivariate functional data}
\usage{
basis_mfd(
  X,
  grid = NULL,
  basis = "fpca",
  FVE = 0.9,
  K = NULL,
  n_basis = 4,
  n_order = 4,
  gram = TRUE
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{grid}{a vector containing m timepoints}

\item{basis}{a choice of the basis. "fpca" (FPCA) or "bspline" (B-spline) is supported.}

\item{FVE}{Fraction of variance explained (Default is 0.90)}

\item{K}{the number of FPCs (Default is selected by FVE)}

\item{n_basis}{the number of basis for the B-spline basis expansion}

\item{n_order}{the order of B-spline basis functions. Default is 4 (cubic B-spline)}

\item{gram}{If TRUE, Gram-Schmidt orthogonalization is performed for the estimated basis coefficients. (Default is TRUE)}
}
\value{
a \code{basis_mfd} object
}
\description{
Make basis expansion matrix from multivariate functional data
}
