% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfda-package.R
\docType{package}
\name{hdfda-package}
\alias{hdfda}
\alias{hdfda-package}
\title{hdfda: High-Dimensional Functional Data Analysis}
\description{
This is the R-package for high-dimensional functional data analysis. It includes the functional regression and classification with sparsity penalty and the statistical inferences.
}
\author{
\strong{Maintainer}: Hyunsung Kim \email{hyunsung1021@gmail.com}

}
\keyword{internal}
