% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfda-package.R
\docType{package}
\name{hdfda-package}
\alias{hdfda}
\alias{hdfda-package}
\title{hdfda: High-Dimensional Functional Data Analysis}
\description{
`hdfda` is the R-package for High-Dimensional Functional Data Analysis. It includes the functional linear regression, linear discriminant analysis and two-sample test for high-dimensional functional data.
}
\author{
\strong{Maintainer}: Hyunsung Kim \email{hyunsung1021@gmail.com}

}
\keyword{internal}
