% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_proj_dir.R
\name{opt_proj_dir}
\alias{opt_proj_dir}
\title{Estimate the Optimal Projection Direction for High-dimensional Functional Data with Group Non-convex Penalty on Two-sample Test problem}
\usage{
opt_proj_dir(
  X,
  y,
  grid = NULL,
  penalty = "scad",
  basis = "bspline",
  n_basis = 4,
  lambda = 0.1,
  a = 3.7,
  max_iter = 1000,
  sweeps_bcd = 2
)
}
\arguments{
\item{X}{a n-m-p array (p-variate functional data; each functional data consists of n curves observed from m timepoints)}

\item{y}{a integer vector containing class label of X (n x 1 vector)}

\item{grid}{a vector containing m timepoints}

\item{penalty}{the penalty type ("scad" is only supported.)}

\item{basis}{Default is "bspline" ("fpca" is possible but takes huge times if p is high)}

\item{n_basis}{the number of cubic B-spline bases using \code{n_basis}-2 knots}

\item{lambda}{a penalty parameter for L1-regularization}

\item{a}{a parameter for SCAD penalty}

\item{max_iter}{a maximum iteration number of the LQA algorithm}

\item{sweeps_bcd}{a number of sweeps in BCD algorithm}
}
\value{
a \code{opt_proj_dir} object
}
\description{
The optimal projection direction is corresponding to the high-dimensional functional linear discriminant direction under the sparsity assumption.
The optimal projection direction is estimated by the block coordinate descent (BCD) algorithm with local quadratic approximation (LQA).
}
\references{
Kim, H. and Park, J. (2026+). Two-sample Projection Test for High-Dimensional Functional Data, Submitted.
}
